tailcfg, hostinfo: put envtype in Hostinfo
